+++ engine/src/main/java/dev/gamekit/utils/Task.java
@@ -1,7 +1,0 @@
-package dev.gamekit.utils;
-
-/** Interface for an arbitrary task */
-public interface Task {
-  /** Abstract method implementing the task's logic */
-  void run();
-}